requirements.txt
streamlit
numpy
Pillow
pandas
geopy
tensorflow==2.6.0